progress

+++ src/components/CanvasEditor/TileGridEdit/data/TileSheetHistory.ts
@@ -11,12 +11,12 @@ export type ProtoTileSheetPatch = {
   y: number
   w: number
   h: number
-  before: Uint8ClampedArray
-  after: Uint8ClampedArray | null
+  before: Uint32Array
+  after: Uint32Array | null
 }
 
 export type TileSheetPatch = Omit<ProtoTileSheetPatch, 'after'> & {
-  after: Uint8ClampedArray
+  after: Uint32Array
 }
 
 export type TileRect = Rect & { tileId: TileId }
